Mejorando validaciones

+++ src/main/java/com/example/demo/categoria/CategoriaServiceImpl.java
@@ -23,12 +23,52 @@ public class CategoriaServiceImpl implements CategoriaService {
     }
 
     @Override
+    public Categoria obtenerPorIdIncluyendoInactivos(String idCategoria) {
+        return categoriaDAO.obtenerPorIdIncluyendoInactivos(idCategoria);
+    }
+
+    @Override
+    public Categoria obtenerPorNombreIncluyendoInactivos(String nombre) {
+        return categoriaDAO.obtenerPorNombreIncluyendoInactivos(nombre);
+    }
+
+    @Override
     public void crearCategoria(Categoria categoria) {
+
+        // Validar ID
+        Categoria idExistente = categoriaDAO.obtenerPorIdIncluyendoInactivos(categoria.getIdCategoria());
+        if (idExistente != null) {
+            if (!idExistente.isEstado()) {
+                throw new IllegalStateException(
+                        "El código ingresado pertenece a una categoría inactiva. Reactívala para volver a usarla.");
+            } else {
+                throw new IllegalStateException(
+                        "El código ingresado ya existe. Usa uno diferente.");
+            }
+        }
+
+        // Validar nombre
+        Categoria nombreExistente = categoriaDAO.obtenerPorNombreIncluyendoInactivos(categoria.getNombre());
+        if (nombreExistente != null) {
+            if (!nombreExistente.isEstado()) {
+                throw new IllegalStateException(
+                        "El nombre ingresado pertenece a una categoría inactiva. Reactívala para volver a usarla.");
+            } else {
+                throw new IllegalStateException(
+                        "El nombre ingresado ya existe. Usa uno diferente.");
+            }
+        }
+
         categoriaDAO.crearCategoria(categoria);
     }
 
     @Override
     public void eliminarCategoria(String idCategoria) {
+        int cantidadProductos = categoriaDAO.contarProductosPorCategoria(idCategoria);
+        if (cantidadProductos > 0) {
+            throw new IllegalStateException("No se puede eliminar la categoría porque tiene productos asociados.");
+        }
+
         categoriaDAO.eliminarCategoria(idCategoria);
     }
